requirements.txt
amqp==5.0.9
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-celery-results==2.2.0
django-import-export==2.7.1
et-xmlfile==1.1.0
kombu==5.2.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
prompt-toolkit==3.0.24
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0

pip~=22.0.3
wheel~=0.36.2
setuptools~=59.6.0
django-crispy-forms~=1.14.0